_unrelated/betting/requirements_apostas.txt
pytesseract==0.3.10
opencv-python==4.8.1.78
Pillow==10.0.1
pyperclip==1.8.2
numpy==1.24.3